\name{DARTData-package}
\alias{DARTData-package}
\alias{DARTData}
\docType{package}
\title{
Example data set for DART package
}
\description{
Data set used in DART package
}
\details{
\tabular{ll}{
Package: \tab DARTData\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-24\cr
License:  \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Andrew E Teschendorff, Yan Jiao

Maintainer:  <y.jiao@ucl.ac.uk>
}

\references{
{Jiao Y, Lawler K, Patel GS, Purushotham A, Jones AF, Grigoriadis A, Ng T, Teschendorff AE. Denoising algorithm based on relevance network topology improves molecular pathway activity inference. Submitted.}

{Teschendorff AE, Gomez S, Arenas A, El-Ashry D, Schmidt M, et al. (2010) Improved prognostic classification of breast
cancer defined by antagonistic activation patterns of immune response pathway modules. BMC Cancer 10:604.}
}

\keyword{ example dataset package }
\seealso{
Wang Y, Klijn JG, Zhang Y, Sieuwerts AM, Look MP, et al. Gene-expression profiles to predict distant metastasis of lymph-node-negative primary breast cancer.  Lancet 365:671-679
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE2034
}
\examples{
### load example data set
data(data.m)
data(sign.v)
}
